Add files via upload
The table shows that the algorithmic trading system has a higher risk-adjusted return compared to the benchmark. The returns of the trading system have lower volatility and a higher Sharpe Ratio compared to the benchmark. Moreover, the trading system has a lower overall risk profile with lower maximum drawdown, Value-at-Risk and Expected Shortfall compared to the SPY benchmark.